requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-huggingface>=0.1.0
streamlit
faiss-cpu
pypdf
scikit-learn
numpy
python-dotenv
transformers>=4.30.0
torch>=2.0.0